Aum_resume_screener_app/requirements.txt
streamlit
sentence-transformers
torch
PyPDF2
python-docx
scikit-learn
pandas
numpy